develop3

+++ lib-routing/src/main/java/ru/geekbrains/market2/routing/dtos/CartDto.java
@@ -3,11 +3,12 @@ package ru.geekbrains.market2.routing.dtos;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @NoArgsConstructor
 @Data
 public class CartDto {
     private List<CartItemDto> items;
-    private int totalPrice;
+    private BigDecimal totalPrice;
 }
